gcpdeploy/requirements.txt
pandas
scikit-learn
mlflow
joblib
flask
google-cloud-storage
gunicorn
mlflow==1.30.0
gcsfs
google-cloud-aiplatform 